Rename backend sources
MSBuild is incredibly brain dead but I don't want to fiddle any more
with the build right now. A .bat file is just little too slow for a full
build. So, add a trivial VS project for now. (Does not build cleanly yet
due to warnings that are cl.exe-only.)