Copy in most of winabi branch from c9x
Not sure if overall a good idea. I think it probably makes more sense to
amd64_sysv with a few `if (T.windows)`, but we'll see how it goes and
maybe it can be merged back once all the tests pass.

Also various fixes for cl to build. There's still a vcproj, but just use
m.bat for now since it's faster. :-(